requirements.txt
anthropic>=0.40.0
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
pytrends>=4.9.2
praw>=7.7.0
